+++ src/agents/workflows/directedTwitter/tools.ts
@@ -1,10 +1,10 @@
 import { TwitterApi } from '../../../services/twitter/types.js';
-import { createFetchMentionsTool } from '../../tools/twitter/fetchMentionsTool.js';
-import { createFetchTimelineTool } from '../../tools/twitter/fetchTimelineTool.js';
-import { createPostTweetTool } from '../../tools/twitter/postTweetTool.js';
-import { createUploadToDsnTool } from '../../tools/uploadToDsnTool.js';
-import { createFetchMyRecentTweetsTool } from '../../tools/twitter/fetchMyRecentTweetsTool.js';
-import { createFetchMyRecentRepliesTool } from '../../tools/twitter/fetchMyRecentRepliesTool.js';
+import { createFetchMentionsTool } from './tools/twitter/fetchMentionsTool.js';
+import { createFetchTimelineTool } from './tools/twitter/fetchTimelineTool.js';
+import { createPostTweetTool } from './tools/twitter/postTweetTool.js';
+import { createUploadToDsnTool } from './tools/uploadToDsnTool.js';
+import { createFetchMyRecentTweetsTool } from './tools/twitter/fetchMyRecentTweetsTool.js';
+import { createFetchMyRecentRepliesTool } from './tools/twitter/fetchMyRecentRepliesTool.js';
 
 export const createTools = (twitterApi: TwitterApi) => {
   const fetchTimelineTool = createFetchTimelineTool(twitterApi);
