+++ src/agents/workflows/directedTwitter/tools/twitter/fetchTimelineTool.ts
@@ -1,7 +1,7 @@
 import { DynamicStructuredTool } from '@langchain/core/tools';
 import { z } from 'zod';
-import { createLogger } from '../../../utils/logger.js';
-import { TwitterApi } from '../../../services/twitter/types.js';
+import { createLogger } from '../../../../../utils/logger.js';
+import { TwitterApi } from '../../../../../services/twitter/types.js';
 import { ToolNode } from '@langchain/langgraph/prebuilt';
 import { AIMessage } from '@langchain/core/messages';
