+++ src/agents/workflows/directedTwitter/tools/uploadToDsnTool.ts
@@ -1,9 +1,9 @@
 import { DynamicStructuredTool } from '@langchain/core/tools';
 import { z } from 'zod';
-import { createLogger } from '../../utils/logger.js';
+import { createLogger } from '../../../../utils/logger.js';
 import { ToolNode } from '@langchain/langgraph/prebuilt';
 import { AIMessage } from '@langchain/core/messages';
-import { uploadToDsn } from './utils/dsn/dsnUpload.js';
+import { uploadToDsn } from '../../../tools/utils/dsn/dsnUpload.js';
 
 const logger = createLogger('upload-to-dsn-tool');
